+++ docs/source/conf.py
@@ -20,13 +20,13 @@
 # -- Project information -----------------------------------------------------
 
 project = 'gr-satellites'
-copyright = '2016-2024, Daniel Estévez'
+copyright = '2016-2025, Daniel Estévez'
 author = 'Daniel Estévez'
 
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '5.7.0'
+release = '5.8.0'
 
 # -- General configuration ---------------------------------------------------