+++ python/core/gr_satellites_flowgraph.py
@@ -272,6 +272,7 @@ class gr_satellites_flowgraph(gr.hier_block2):
             'RS interleaving': 'rs_interleaving',
             'convolutional': 'convolutional',
             'scrambler': 'scrambler',
+            'callsign': 'callsign',
             }
         deframer_additional_options = filter_translate_dict(transmitter,
                                                             deframe_options)
@@ -452,6 +453,10 @@ class gr_satellites_flowgraph(gr.hier_block2):
         'UA01': deframers.ua01_deframer,
         'Mobitex': deframers.mobitex_deframer,
         'Mobitex-NX': set_options(deframers.mobitex_deframer, nx=True),
+        'BEESAT-1': set_options(deframers.mobitex_deframer, nx=True,
+                                variant='BEESAT-1'),
+        'BEESAT-9': set_options(deframers.mobitex_deframer, nx=True,
+                                variant='BEESAT-9'),
         'FOSSASAT': deframers.fossasat_deframer,
         'AISTECHSAT-2': deframers.aistechsat_2_deframer,
         'AALTO-1': deframers.aalto1_deframer,