+++ python/telemetry/__init__.py
@@ -48,6 +48,7 @@ from .kr01 import kr01
 from .lume import lume
 from .mirsat1 import mirsat1
 from .mysat1 import mysat1
+from .nila import nila
 from .picsat import picsat
 from .qo100 import qo100
 from .quetzal1 import quetzal1
@@ -61,6 +62,7 @@ from .snet import snet
 from .suomi100 import suomi100
 from .by70_1 import taurus1
 from .trisat import trisat
+from .tubix20 import tubix20
 from .upmsat_2 import upmsat_2
 from .vzlusat_2 import vzlusat_2